Remove custom geometry structures and migrate to SkiaSharp types.

+++ src/FloatSoda.Common/Geometries/Offset.cs
@@ -1,6 +1,6 @@
 ﻿namespace FloatSoda.Common.Geometries;
 
-public readonly record struct Offset(float X, float Y)
+public readonly record struct Offset(double X, double Y)
 {
     public static Offset Zero => new(0, 0);
     public static Offset One => new(1, 1);
@@ -9,8 +9,8 @@ public readonly record struct Offset(float X, float Y)
 
     public static Offset operator -(Offset a, Offset b) => new(a.X - b.X, a.Y - b.Y);
 
-    public static Offset operator *(Offset a, float b) => new(a.X * b, a.Y * b);
-    public static Offset operator *(float a, Offset b) => b * a;
+    public static Offset operator *(Offset a, double b) => new(a.X * b, a.Y * b);
+    public static Offset operator *(double a, Offset b) => b * a;
 
-    public static Offset operator /(Offset a, float b) => new(a.X / b, a.Y / b);
+    public static Offset operator /(Offset a, double b) => new(a.X / b, a.Y / b);
 }
